Add -wo to Little Women loop
If you do not add the -w to the grep command in this loop, it will give you false results, It will return results for all the strings with Jo, Meg, Beth, and Amy in them. This is particularly glaring with "Jo" since it picks up John and other strings with Jo in them. I added the -o to the loop as that will pick up if the string if it appears more than once in a line.  I changed the results to reflect what I got when I ran this loop.