Update 02-navigating-the-filesystem.md
Edited the first bash command

I moved the colon to the end of the parenthetical and added the words "Note that" at the start of the parenthetical to clarify what is going on here. Also, the colon next to the opening parenthesis looked like a frowning face, so moving it to the end is more visually appealing.