Avoid problem with active bang in zsh
MacOS Catalina uses zsh by default, and for zsh `!` is an active character that swallows the following double quote character (in a double-quoted string). By changing the example to use a single-quoted string, the `!` is not active and the example becomes portable between zsh and bash. See issues /issues/96 and /issues/115.