Counting number of files part I and II exercises are a bit misleading and probably unnecessary
The Counting number of files part I and II exercises in the "Counting and Mining in the Shell" episode can be easily done by the command ls | wc -l since the total is not included in an ls command, so I amended that excercise to reflect this. Since the lesson returns to this exercise later in the episode to fix this "one more" problem that no longer exists, I got rid of that grep challenge. Another grep challenge using regular expressions and the -v flag could be added here. This was pointed out to me by a student the last time I taught this lesson.